requirements.txt
alabaster==0.7.9
appdirs==1.4.2
Babel==2.3.4
beautifulsoup4==4.5.3
bottle==0.12.7
bs4==0.0.1
cffi==1.9.1
Cheetah==2.4.4
CherryPy==3.6.0
configobj==5.0.6
CouchDB==1.0
cryptography==1.7.2
cssselect==0.9.1
cycler==0.10.0
decorator==3.4.2
Django==1.11.1
django-classy-tags==0.8.0
django-cms==3.1.2
django-debug-toolbar==1.8
django-debug-toolbar-mongo==0.1.10
django-filter==1.0.4
django-haystack==2.6.1
django-mongodb-engine==0.6.0
django-mongodbforms==0.3.1
django-sekizai==0.8.2
django-treebeard==3.0
djangocms-admin-style==0.2.7
djangorestframework==3.6.3
djangotoolbox==1.8.0
docutils==0.13.1
enum34==1.1.6
evernote==1.25.2
Flask==0.12.1
functools32==3.2.3.post2
gunicorn==19.3.0
haystack==0.15
html2text==2016.9.19
html5lib==0.999
httplib2==0.10.3
idna==2.2
imagesize==0.7.1
ipaddr==2.1.10
ipaddress==1.0.18
Jinja2==2.9.6
keyring==5.7
lxml==3.4.4
Markdown==2.6.2
markdown2==2.3.3
MarkupSafe==1.0
matplotlib==2.0.0
mock==1.0.1
mongoengine==0.13.0
nose==1.3.7
not==2.0.2
numpy==1.10.4
oauth2==1.9.0.post1
packaging==16.8
pdfrw==0.1
Pillow==3.0.0
psycopg2==2.6.1
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pycups==1.9.72
pycurl==7.19.5.1
Pygments==2.2.0
pymongo==3.4.0
pyOpenSSL==0.15.1
pyparsing==2.1.10
pyserial==2.7
pysolr==3.6.0
pytest==3.0.7
pytest-django==3.1.2
python-dateutil==2.6.0
python-ptrace==0.8.1
pytz==2017.2
pyudev==0.17
pyusb==1.0.0b1
PyXB==1.2.4
rebar==0.1
reportlab==2.7
requests==2.14.2
scipy==0.16.1
selenium==2.44.0
simplejson==3.6.5
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
sphinx-rtd-theme==0.1.8
sqlparse==0.2.3
subprocess32==3.2.7
svg2rlg==0.3
svn==0.3.44
Twisted==15.5.0
url==0.4.0
Werkzeug==0.12.1
xmltodict==0.9.2
zope.interface==4.1.2
